Add maximize and sizecompare libs
