+++ src/main.rs
@@ -1,9 +1,11 @@
 // Import modules
-use niri_ipc::{Event, Output, Request, state::EventStreamState, state::EventStreamStatePart};
+use niri_ipc::{Event, state::EventStreamState, state::EventStreamStatePart};
 use std::collections::HashMap;
 
 // Import internal libraries
+mod maximize;
 mod outputs; // https://github.com/Antiz96/oniri/issues/3
+mod sizecompare; // https://github.com/Antiz96/oniri/issues/3
 mod socket;
 mod version;
 
@@ -24,139 +26,60 @@ fn main() -> anyhow::Result<()> {
     // Read events gathered from the IPC socket
     let mut read_event = event_socket.read_events();
 
-    // Create a window(s)/workspace map
+    // Create a workspace/window(s) map
     let mut workspace_windows: HashMap<u64, Vec<u64>> = HashMap::new();
 
-    // Loop over events and filter the ones about windows being opened or closed
-    // Update the window(s)/workspace map when events are matched
+    // Loop over events
     while let Ok(event) = read_event() {
         // Update state
         // This can be dropped once https://github.com/Antiz96/oniri/issues/3 is resolved
         state.apply(event.clone());
 
+       // Filter events
         match event {
+            // Window being opened
             Event::WindowOpenedOrChanged { window } => {
                 // Skip floating windows (they cannot/should not be maximized)
                 if window.is_floating {
                     continue;
                 }
+
                 println!("Trigger Event: Window Opened");
+
+                // Update the workspace/window(s) map
                 let id = window.id;
                 if let Some(ws) = window.workspace_id {
                     workspace_windows.entry(ws).or_default().push(id);
                 }
-                // Maximize the window if it's the only one
-                for windows in workspace_windows.values() {
-                    if windows.len() == 1 {
-                        let id = windows[0];
-                        // https://github.com/Antiz96/oniri/issues/3
-                        if !is_maximized(&state, &outputs, id) {
-                            let _ = action_socket
-                                .send(Request::Action(niri_ipc::Action::MaximizeColumn {}));
-                            println!("Maximized window {}", id);
-                        }
-                    }
-                }
+
+                // Check if there's only one window in the workspace/window(s) map & maximize it if so
+                maximize::maximize_window_if_alone(
+                    &workspace_windows,
+                    &state, // https://github.com/Antiz96/oniri/issues/3
+                    &outputs, // https://github.com/Antiz96/oniri/issues/3
+                    &mut action_socket,
+                )?;
             }
+            // Window being closed
             Event::WindowClosed { id } => {
                 println!("Trigger Event: Window Closed");
+
+                // Update the workspace/window(s) map
                 for windows in workspace_windows.values_mut() {
                     windows.retain(|&wid| wid != id);
                 }
-                // Maximize the window if it's the only one
-                for windows in workspace_windows.values() {
-                    if windows.len() == 1 {
-                        let id = windows[0];
-                        // https://github.com/Antiz96/oniri/issues/3
-                        if !is_maximized(&state, &outputs, id) {
-                            let _ = action_socket
-                                .send(Request::Action(niri_ipc::Action::MaximizeColumn {}));
-                            println!("Maximized window {}", id);
-                        }
-                    }
-                }
+
+                // Check if there's only one window in the workspace/window(s) map & maximize it if so
+                maximize::maximize_window_if_alone(
+                    &workspace_windows,
+                    &state, // https://github.com/Antiz96/oniri/issues/3
+                    &outputs, // https://github.com/Antiz96/oniri/issues/3
+                    &mut action_socket,
+                )?;
             }
             // Ignore other events
             _ => {}
         }
     }
     Ok(())
-}
-
-// Hack to determine if a window is supposedly already maximized or not
-// based on a comparison between the window size and the output size.
-// This is to workaround the lack of a window "maximized" state to gather from the IPC,
-// and/or the lack of a way to set/unset the maximize state (rather than just toggling it).
-// See https://github.com/Antiz96/oniri/issues/3 for more details.
-fn is_maximized(
-    state: &EventStreamState,
-    outputs: &HashMap<String, Output>,
-    window_id: u64,
-) -> bool {
-    let tol_w = 150;
-    let tol_h = 150;
-
-    let window = match state.windows.windows.get(&window_id) {
-        Some(w) => w,
-        None => {
-            println!("Window {} not found in state", window_id);
-            return false;
-        }
-    };
-
-    if window.is_floating {
-        println!("Window {} is floating, skipping", window_id);
-        return false;
-    }
-
-    let workspace = match window
-        .workspace_id
-        .and_then(|ws_id| state.workspaces.workspaces.get(&ws_id))
-    {
-        Some(ws) => ws,
-        None => {
-            println!("Workspace for window {} not found", window_id);
-            return false;
-        }
-    };
-
-    let logical = match workspace
-        .output
-        .as_ref()
-        .and_then(|name| outputs.get(name))
-        .and_then(|o| o.logical.as_ref())
-    {
-        Some(l) => l,
-        None => {
-            println!(
-                "Output for workspace {} not found or has no logical size",
-                workspace.id
-            );
-            return false;
-        }
-    };
-
-    let out_w = logical.width as i32;
-    let out_h = logical.height as i32;
-    let (tile_w, tile_h) = window.layout.tile_size;
-    let tile_w = tile_w as i32;
-    let tile_h = tile_h as i32;
-
-    println!(
-        "Window {}: out_w={}, out_h={}, tile_w={}, tile_h={}, tol_w={}, tol_h={}",
-        window_id, out_w, out_h, tile_w, tile_h, tol_w, tol_h
-    );
-
-    let width_ok = (out_w - tile_w).abs() <= tol_w;
-    let height_ok = (out_h - tile_h).abs() <= tol_h;
-
-    println!(
-        "Window {}: width_ok={}, height_ok={}, maximized={}",
-        window_id,
-        width_ok,
-        height_ok,
-        width_ok && height_ok
-    );
-
-    width_ok && height_ok
 }
