+++ src/maximize.rs
@@ -1,0 +1,27 @@
+// Import modules
+use std::collections::HashMap;
+use niri_ipc::{Request, socket::Socket, Output};
+
+// Import internal libraries
+use crate::sizecompare::is_maximized; // https://github.com/Antiz96/oniri/issues/3
+
+// Check if there's only one window in the workspace/window(s) map & maximize it if so,
+// unless it's maximized already (https://github.com/Antiz96/oniri/issues/3)
+pub fn maximize_window_if_alone(
+    workspace_windows: &HashMap<u64, Vec<u64>>,
+    state: &niri_ipc::state::EventStreamState, // https://github.com/Antiz96/oniri/issues/3
+    outputs: &HashMap<String, Output>, // https://github.com/Antiz96/oniri/issues/3
+    action_socket: &mut Socket,
+) -> anyhow::Result<()> {
+    for windows in workspace_windows.values() {
+        if windows.len() == 1 {
+            let id = windows[0];
+            if !is_maximized(&state, &outputs, id) { // https://github.com/Antiz96/oniri/issues/3
+                let _ = action_socket
+                    .send(Request::Action(niri_ipc::Action::MaximizeColumn {}));
+                println!("Maximized window {}", id);
+            }
+        }
+    }
+    Ok(())
+}
